Sépare la logique du rendu d'email dans un controller
